new foraging-based giveup logic in place and cleaned up PVDA logic; added Time in addition to Effort; minimally working but params need adjusting

+++ axon/pcore_layers.go
@@ -165,7 +165,7 @@ func (ly *Layer) MatrixGated(ctx *Context) {
 		}
 		if ctx.PlusPhase.IsTrue() && ly.Params.Matrix.IsVS.IsTrue() {
 			SetGlbV(ctx, di, GvVSMatrixJustGated, bools.ToFloat32(mtxGated))
-			SetGlbUSposV(ctx, di, GvVSMatrixPoolGated, poolIdx, 1)
+			SetGlbUSposV(ctx, di, GvVSMatrixPoolGated, uint32(poolIdx), 1)
 		}
 	}
 }
